+++ Cargo.toml
@@ -1,2 +1,16 @@
 [workspace]
-members = ["askama", "askama_derive", "askama_escape", "askama_shared", "testing"]
+members = [
+    "askama",
+    "askama_actix",
+    "askama_derive",
+    "askama_escape",
+    "askama_shared",
+    "testing",
+]
+default-members = [
+    "askama",
+    "askama_derive",
+    "askama_escape",
+    "askama_shared",
+    "testing",
+]
