+++ askama/Cargo.toml
@@ -26,7 +26,7 @@ serde-yaml = ["askama_shared/yaml"]
 num-traits = ["askama_shared/num-traits"]
 with-iron = ["iron", "askama_derive/iron"]
 with-rocket = ["rocket", "askama_derive/rocket"]
-with-actix-web = ["actix-web", "askama_derive/actix-web", "mime", "mime_guess", "bytes", "futures"]
+with-actix-web = ["askama_derive/actix-web"]
 with-gotham = ["gotham", "askama_derive/gotham", "hyper", "mime", "mime_guess"]
 
 [dependencies]
@@ -35,13 +35,10 @@ askama_escape = { version = "0.3.0", path = "../askama_escape" }
 askama_shared = { version = "0.9.0", path = "../askama_shared", default-features = false }
 iron = { version = ">= 0.5, < 0.7", optional = true }
 rocket = { version = "0.4", default-features = false, optional = true }
-actix-web = { version = "2", optional = true }
-futures = { version = "0.3", optional = true }
 mime = { version = "0.3", optional = true }
 mime_guess = { version = "2.0.0-alpha", optional = true }
 gotham = { version = "0.3", default-features = false, optional = true }
 hyper = { version = "0.12", optional = true }
-bytes = { version = "0.5", optional = true }
 
 [package.metadata.docs.rs]
 features = ["config", "humansize", "num-traits", "serde-json", "serde-yaml"]
