+++ askama_actix/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "askama_actix"
+version = "0.9.0"
+authors = ["Dirkjan Ochtman <dirkjan@ochtman.nl>"]
+description = "Actix-Web integration for Askama templates"
+documentation = "https://docs.rs/askama"
+keywords = ["markup", "template", "jinja2", "html"]
+categories = ["template-engine"]
+homepage = "https://github.com/djc/askama"
+repository = "https://github.com/djc/askama"
+license = "MIT OR Apache-2.0"
+workspace = ".."
+edition = "2018"
+
+[dependencies]
+actix-web = { version = "2" }
+askama = { version = "0.9.0", path = "../askama", features = ["with-actix-web", "mime", "mime_guess"] }
+bytes = { version = "0.5" }
+futures = { version = "0.3" }
+
+[dev-dependencies]
+actix-rt = { version = "1", default-features = false }
