+++ askama_actix/templates/hello.html
@@ -1,0 +1,1 @@
+Hello, {{ name }}!
