+++ askama_derive/src/generator.rs
@@ -241,15 +241,15 @@ impl<'a> Generator<'a> {
 
     // Implement Actix-web's `Responder`.
     fn impl_actix_web_responder(&mut self, buf: &mut Buffer) {
-        self.write_header(buf, "::askama::actix_web::Responder", None);
-        buf.writeln("type Future = ::futures::future::Ready<::std::result::Result<::askama::actix_web::HttpResponse, Self::Error>>;");
-        buf.writeln("type Error = ::askama::actix_web::Error;");
+        self.write_header(buf, "::actix_web::Responder", None);
+        buf.writeln("type Future = ::futures::future::Ready<::std::result::Result<::actix_web::HttpResponse, Self::Error>>;");
+        buf.writeln("type Error = ::actix_web::Error;");
         buf.writeln(
-            "fn respond_to(self, _req: &::askama::actix_web::HttpRequest) \
+            "fn respond_to(self, _req: &::actix_web::HttpRequest) \
              -> Self::Future {",
         );
 
-        buf.writeln("use ::askama::actix_web::TemplateIntoResponse;");
+        buf.writeln("use ::askama_actix::TemplateIntoResponse;");
         buf.writeln("::futures::future::ready(self.into_response())");
 
         buf.writeln("}");
