+++ testing/Cargo.toml
@@ -6,20 +6,15 @@ workspace = ".."
 edition = "2018"
 
 [features]
-actix = ["actix-web", "actix-rt", "bytes", "askama/with-actix-web", "futures"]
 default = []
-full = ["actix", "with-iron", "serde-json", "with-gotham"]
+full = ["with-iron", "serde-json", "with-gotham"]
 serde-json = ["serde_json", "askama/serde-json"]
 with-rocket = ["rocket", "askama/with-rocket"]
 with-iron = ["iron", "askama/with-iron"]
 with-gotham = ["gotham", "askama/with-gotham", "mime", "hyper"]
 
 [dependencies]
-actix-web = { version = "2", optional = true }
-actix-rt = { version = "1", optional = true }
-futures = { version = "0.3", optional = true }
 askama = { path = "../askama", version = "*" }
-bytes = { version = "0.5", optional = true }
 iron = { version = "0.6", optional = true }
 rocket = { version = "0.4", default-features = false, optional = true }
 serde_json = { version = "1.0", optional = true }
